Made the registered side channels a static property and created the sideChannelUtils class to handle side channel stuff
