+++ Project/Assets/ML-Agents/Examples/3DBall/Scripts/Ball3DAgent.cs
@@ -13,7 +13,7 @@ public class Ball3DAgent : Agent
     public override void Initialize()
     {
         m_BallRb = ball.GetComponent<Rigidbody>();
-        m_ResetParams = Academy.Instance.GetSideChannel<FloatPropertiesChannel>();
+        m_ResetParams = SideChannelUtils.GetSideChannel<FloatPropertiesChannel>();
         SetResetParameters();
     }
