+++ Project/Assets/ML-Agents/Examples/GridWorld/Scripts/GridSettings.cs
@@ -8,7 +8,7 @@ public class GridSettings : MonoBehaviour
 
     public void Awake()
     {
-        Academy.Instance.GetSideChannel<FloatPropertiesChannel>().RegisterCallback("gridSize", f =>
+        SideChannelUtils.GetSideChannel<FloatPropertiesChannel>().RegisterCallback("gridSize", f =>
         {
             MainCamera.transform.position = new Vector3(-(f - 1) / 2f, f * 1.25f, -(f - 1) / 2f);
             MainCamera.orthographicSize = (f + 5f) / 2f;
