+++ Project/Assets/ML-Agents/Examples/Reacher/Scripts/ReacherAgent.cs
@@ -110,7 +110,7 @@ public class ReacherAgent : Agent
 
     public void SetResetParameters()
     {
-        var fp = Academy.Instance.GetSideChannel<FloatPropertiesChannel>();
+        var fp = SideChannelUtils.GetSideChannel<FloatPropertiesChannel>();
         m_GoalSize = fp.GetPropertyWithDefault("goal_size", 5);
         m_GoalSpeed = Random.Range(-1f, 1f) * fp.GetPropertyWithDefault("goal_speed", 1);
         m_Deviation = fp.GetPropertyWithDefault("deviation", 0);
