+++ com.unity.ml-agents/Runtime/Academy.cs
@@ -86,13 +86,6 @@ namespace MLAgents
         /// </summary>
         public static Academy Instance { get { return s_Lazy.Value; } }
 
-        /// <summary>
-        /// Collection of float properties (indexed by a string).
-        /// </summary>
-        [Obsolete("Academy.FloatProperties has been deprecated, use Academy.GetSideChannel<>() instead.")]
-        public FloatPropertiesChannel FloatProperties;
-
-
         // Fields not provided in the Inspector.
 
         /// <summary>
@@ -216,54 +209,6 @@ namespace MLAgents
         }
 
         /// <summary>
-        /// Registers SideChannel to the Academy to send and receive data with Python.
-        /// If IsCommunicatorOn is false, the SideChannel will not be registered.
-        /// </summary>
-        /// <param name="channel"> The side channel to be registered.</param>
-        public void RegisterSideChannel(SideChannel channel)
-        {
-            LazyInitialize();
-            Communicator?.RegisterSideChannel(channel);
-        }
-
-        /// <summary>
-        /// Unregisters SideChannel to the Academy. If the side channel was not registered,
-        /// nothing will happen.
-        /// </summary>
-        /// <param name="channel"> The side channel to be unregistered.</param>
-        public void UnregisterSideChannel(SideChannel channel)
-        {
-            Communicator?.UnregisterSideChannel(channel);
-        }
-
-        /// <summary>
-        /// Returns the SideChannel of Type T if there is one registered, or null if it doesn't.
-        /// If there are multiple SideChannels of the same type registered, the returned instance is arbitrary.
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <returns></returns>
-        public T GetSideChannel<T>() where T: SideChannel
-        {
-            return Communicator?.GetSideChannel<T>();
-        }
-
-        /// <summary>
-        /// Returns all SideChannels of Type T that are registered. Use <see cref="GetSideChannel{T}()"/> if possible,
-        /// as that does not make any memory allocations.
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <returns></returns>
-        public List<T> GetSideChannels<T>() where T: SideChannel
-        {
-            if (Communicator == null)
-            {
-                // Make sure we return a non-null List.
-                return new List<T>();
-            }
-            return Communicator.GetSideChannels<T>();
-        }
-
-        /// <summary>
         /// Disable stepping of the Academy during the FixedUpdate phase. If this is called, the Academy must be
         /// stepped manually by the user by calling Academy.EnvironmentStep().
         /// </summary>
@@ -343,10 +288,9 @@ namespace MLAgents
         {
             EnableAutomaticStepping();
 
-            var floatProperties = new FloatPropertiesChannel();
-#pragma warning disable 0618
-            FloatProperties = floatProperties;
-#pragma warning restore 0618
+            SideChannelUtils.RegisterSideChannel(new EngineConfigurationChannel());
+            SideChannelUtils.RegisterSideChannel(new FloatPropertiesChannel());
+            SideChannelUtils.RegisterSideChannel(new StatsSideChannel());
 
             // Try to launch the communicator by using the arguments passed at launch
             var port = ReadPortFromArgs();
@@ -362,9 +306,6 @@ namespace MLAgents
 
             if (Communicator != null)
             {
-                Communicator.RegisterSideChannel(new EngineConfigurationChannel());
-                Communicator.RegisterSideChannel(floatProperties);
-                Communicator.RegisterSideChannel(new StatsSideChannel());
                 // We try to exchange the first message with Python. If this fails, it means
                 // no Python Process is ready to train the environment. In this case, the
                 //environment must use Inference.
@@ -549,6 +490,7 @@ namespace MLAgents
 
             Communicator?.Dispose();
             Communicator = null;
+            SideChannelUtils.UnregisterAllSideChannels();
 
             if (m_ModelRunners != null)
             {
@@ -566,9 +508,6 @@ namespace MLAgents
             // TODO - Pass worker ID or some other identifier,
             // so that multiple envs won't overwrite each others stats.
             TimerStack.Instance.SaveJsonTimers();
-#pragma warning disable 0618
-            FloatProperties = null;
-#pragma warning restore 0618
             m_Initialized = false;
 
             // Reset the Lazy instance
