+++ com.unity.ml-agents/Runtime/SideChannels/SideChannelUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2506dff31271f49298fbff21e13fa8b6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
