+++ docs/Custom-SideChannels.md
@@ -24,7 +24,7 @@ To send data from C# to Python, create an `OutgoingMessage` instance, add data t
 `base.QueueMessageToSend(msg)` method inside the side channel, and call the
 `OutgoingMessage.Dispose()` method.
 
-To register a side channel on the Unity side, call `Academy.Instance.RegisterSideChannel` with the side channel
+To register a side channel on the Unity side, call `SideChannelUtils.RegisterSideChannel` with the side channel
 as only argument.
 
 ### Python side
@@ -123,7 +123,7 @@ public class RegisterStringLogSideChannel : MonoBehaviour
         Application.logMessageReceived += stringChannel.SendDebugStatementToPython;
 
         // The channel must be registered with the Academy
-        Academy.Instance.RegisterSideChannel(stringChannel);
+        SideChannelUtils.RegisterSideChannel(stringChannel);
     }
 
     public void OnDestroy()
@@ -131,7 +131,7 @@ public class RegisterStringLogSideChannel : MonoBehaviour
         // De-register the Debug.Log callback
         Application.logMessageReceived -= stringChannel.SendDebugStatementToPython;
         if (Academy.IsInitialized){
-            Academy.Instance.UnregisterSideChannel(stringChannel);
+            SideChannelUtils.UnregisterSideChannel(stringChannel);
         }
     }
