+++ docs/Migrating.md
@@ -10,10 +10,14 @@ The versions can be found in
 ## Migrating from 0.15 to latest
 
 ### Important changes
-* `Academy.FloatProperties` was deprecated. It will be removed in a later release.
+* `Academy.FloatProperties` was removed.
+* `Academy.RegisterSideChannel` and `Academy.UnregisterSideChannel` were removed.
+
 
 ### Steps to Migrate
-* Replace `Academy.FloatProperties` with `Academy.GetSideChannel<FloatPropertiesChannel>()`.
+* Replace `Academy.FloatProperties` with `SideChannelUtils.GetSideChannel<FloatPropertiesChannel>()`.
+* Replace `Academy.RegisterSideChannel` with `SideChannelUtils.RegisterSideChannel()`.
+* Replace `Academy.UnregisterSideChannel` with `SideChannelUtils.UnregisterSideChannel`.
 
 
 ## Migrating from 0.14 to 0.15
